Add local S3 server test rule.
