+++ src/test/java/tdl/s3/LocalS3Test.java
@@ -5,27 +5,39 @@ import com.amazonaws.auth.AnonymousAWSCredentials;
 import com.amazonaws.client.builder.AwsClientBuilder.EndpointConfiguration;
 import com.amazonaws.services.s3.AmazonS3;
 import com.amazonaws.services.s3.AmazonS3ClientBuilder;
-import io.findify.s3mock.S3Mock;
+import org.junit.ClassRule;
 import org.junit.Rule;
 import org.junit.Test;
+import org.junit.rules.RuleChain;
 import tdl.s3.rules.LocalS3Server;
+import tdl.s3.rules.LocalTestBucket;
 import tdl.s3.rules.RemoteTestBucket;
 
 public class LocalS3Test {
 
+    @ClassRule
+    public static LocalS3Server localS3 = new LocalS3Server();
+
     @Rule
-    public LocalS3Server localS3 = new LocalS3Server();
+    public LocalTestBucket testBucket = new LocalTestBucket();
 
     @Test
-    public void run() {
-        EndpointConfiguration endpoint = new EndpointConfiguration("http://localhost:8001", "us-west-2");
-        AmazonS3 client = AmazonS3ClientBuilder
-                .standard()
-                .withPathStyleAccessEnabled(true)
-                .withCredentials(new AWSStaticCredentialsProvider(new AnonymousAWSCredentials()))
-                .withEndpointConfiguration(endpoint)
-                .build();
+    public void run1() {
+        AmazonS3 client = testBucket.getAmazonS3();
+        client.createBucket("testbucket");
+        client.putObject("testbucket", "file/name", "contents");
+    }
 
+    @Test
+    public void run2() {
+        AmazonS3 client = testBucket.getAmazonS3();
+        client.createBucket("testbucket");
+        client.putObject("testbucket", "file/name", "contents");
+    }
+
+    @Test
+    public void run3() {
+        AmazonS3 client = testBucket.getAmazonS3();
         client.createBucket("testbucket");
         client.putObject("testbucket", "file/name", "contents");
     }
